Move KD Tree implementation to header file
